refactor: divide advert photos and favorite services; reorganize utilities and mappers
