+++ src/adverts/adverts.service.ts
@@ -1,77 +1,23 @@
-import {
-  BadRequestException,
-  ForbiddenException,
-  Injectable,
-  UnauthorizedException,
-} from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { Category, Prisma, Status } from '@prisma/client';
 import { CreateAdvertDto } from './dto/create-advert.dto';
 import { UpdateAdvertDto } from './dto/update-advert.dto';
-import { AdvertEntity } from './entities/advert.entity';
-import { assertOwnership } from 'src/utils/assert-ownership.util';
-import { parseEnumValue } from 'src/utils/parse-enum-value.util';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { PhotoEntity } from './entities/photo.entity';
-import { FilterAdvertDto } from './dto/filter-advert.dto';
 import {
-  OwnerDetailsEntity,
-  OwnerSummaryEntity,
-} from 'src/users/entities/owner.entity';
+  assertOwnership,
+  getPagination,
+  resolveCoordinates,
+  parseEnumValue,
+} from './adverts.helper';
+import { PrismaService } from 'src/prisma/prisma.service';
+import { FilterAdvertDto } from './dto/filter-advert.dto';
+import { AdvertsMapper } from './adverts.mapper';
 
 @Injectable()
 export class AdvertsService {
-  constructor(private prisma: PrismaService) {}
-
-  /**
-   * Calculates pagination parameters (limit and offset) from filter DTO.
-   * @param filters - DTO containing filter pagination options
-   * @returns Object with `limit` and `offset` values for SQL queries.
-   */
-  private getPagination(filters: FilterAdvertDto) {
-    const { limit = 20, page = 1 } = filters;
-
-    return { limit, offset: (page - 1) * limit };
-  }
-
-  /**
-   * Resolves the geographic coordinates to be used for distance calculations.
-   * Priority order:
-   * 1. Logged user's location (if `useUserLocation` is `true`).
-   * 2. Manual coordinates provided in filters.
-   * 3. No coordinates (distance filtering disabled).
-   * @param userId - ID of the authenticated user (nullable)
-   * @param filters - Advert filter DTO
-   * @throws UnauthorizedException if user location is requested but no user is logged in.
-   * @throws BaqRequestException if user has no defined location.
-   * @returns Coordinates object or `null` if no location should be used.
-   */
-  private async resolveCoordinates(
-    userId: string | null,
-    filters: FilterAdvertDto,
-  ): Promise<{ latitude: number; longitude: number } | null> {
-    if (filters.useUserLocation) {
-      if (!userId) {
-        throw new UnauthorizedException('Login required to use user location');
-      }
-
-      const user = await this.prisma.user.findUnique({
-        where: { id: userId },
-        select: { latitude: true, longitude: true },
-      });
-
-      if (user?.latitude == null || user?.longitude == null) {
-        throw new BadRequestException('User has no location set');
-      }
-
-      return { latitude: user.latitude, longitude: user.longitude };
-    }
-
-    if (filters.latitude != null && filters.longitude != null) {
-      return { latitude: filters.latitude, longitude: filters.longitude };
-    }
-
-    return null;
-  }
+  constructor(
+    private prisma: PrismaService,
+    private readonly mapper: AdvertsMapper,
+  ) {}
 
   /**
    * Builds a SQL expression to calculate distance (in Km).
@@ -282,19 +228,9 @@ export class AdvertsService {
       },
     });
 
-    return new AdvertEntity({
-      ...createdAdvert,
-      photos: createdAdvert.photos.map((photo) => new PhotoEntity(photo)),
-      owner: new OwnerDetailsEntity({
-        id: createdAdvert.owner.id,
-        username: createdAdvert.owner.username,
-        postalCode: createdAdvert.owner.postalCode,
-        createdAt: createdAdvert.owner.createdAt,
-        advertsCount: createdAdvert.owner._count.adverts,
-      }),
+    return this.mapper.toAdvertDetailsEntity(createdAdvert, {
       isOwner: true,
       isFavorite: false,
-      favoriteCount: createdAdvert._count.favorites,
     });
   }
 
@@ -307,8 +243,8 @@ export class AdvertsService {
    * @throws UnauthorizedException if user location is requested without authentication.
    */
   async getAll(userId: string | null, filters: FilterAdvertDto) {
-    const { limit, offset } = this.getPagination(filters);
-    const coordinates = await this.resolveCoordinates(userId, filters);
+    const { limit, offset } = getPagination(filters);
+    const coordinates = await resolveCoordinates(userId, filters);
     const { whereSql, distanceSql, distanceFilterSql } =
       this.buildWhereConditions(filters, coordinates);
 
@@ -322,22 +258,14 @@ export class AdvertsService {
       popular: filters.popular ?? false,
     });
 
-    return adverts.map((advert) => {
-      new AdvertEntity({
-        ...advert,
-        owner: new OwnerSummaryEntity({
-          id: advert.ownerId,
-          username: advert.ownerUsername,
-        }),
-        photos: advert.thumbnailUrl
-          ? [new PhotoEntity({ url: advert.thumbnail })]
-          : [],
+    return adverts.map((advert) =>
+      this.mapper.toAdvertSummaryEntity(advert, {
         isOwner: advert.ownerId === userId,
         isFavorite: advert.isFavorite,
-        favoriteCount: Number(advert.favoriteCount),
+        favoriteCount: Number(advert.favoriteCount) ?? 0,
         distance: advert.distance ?? null,
-      });
-    });
+      }),
+    );
   }
 
   /**
@@ -367,10 +295,10 @@ export class AdvertsService {
 
     const isOwner = userId === advert.ownerId;
     const isFavorite = userId
-      ? await this.prisma.favorite.findUnique({
+      ? (await this.prisma.favorite.findUnique({
           where: { userId_advertId: { userId, advertId: advert.id } },
-        })
-      : null;
+        })) !== null
+      : false;
 
     if (!isOwner) {
       await this.prisma.advert.update({
@@ -381,19 +309,9 @@ export class AdvertsService {
       advert.views += 1;
     }
 
-    return new AdvertEntity({
-      ...advert,
-      owner: new OwnerDetailsEntity({
-        id: advert.owner.id,
-        username: advert.owner.username,
-        postalCode: advert.owner.postalCode,
-        createdAt: advert.owner.createdAt,
-        advertsCount: advert.owner._count.adverts,
-      }),
-      photos: advert.photos.map((photo) => new PhotoEntity(photo)),
+    return this.mapper.toAdvertDetailsEntity(advert, {
       isOwner,
-      isFavorite: !!isFavorite,
-      favoriteCount: advert._count.favorites,
+      isFavorite,
     });
   }
 
@@ -413,7 +331,6 @@ export class AdvertsService {
 
     assertOwnership(advert, userId);
 
-    // Normalize enums if provided
     if (updateDto.category) {
       updateAdvertDto.category = parseEnumValue(updateDto.category, Category);
     }
@@ -440,19 +357,9 @@ export class AdvertsService {
       },
     });
 
-    return new AdvertEntity({
-      ...updatedAdvert,
-      owner: new OwnerDetailsEntity({
-        id: updatedAdvert.owner.id,
-        username: updatedAdvert.owner.username,
-        postalCode: updatedAdvert.owner.postalCode,
-        createdAt: updatedAdvert.owner.createdAt,
-        advertsCount: updatedAdvert.owner._count.adverts,
-      }),
-      photos: updatedAdvert.photos.map((photo) => new PhotoEntity(photo)),
+    return this.mapper.toAdvertDetailsEntity(updatedAdvert, {
       isOwner: true,
       isFavorite: false,
-      favoriteCount: updatedAdvert._count.favorites,
     });
   }
 
@@ -485,19 +392,9 @@ export class AdvertsService {
 
     await this.prisma.advert.delete({ where: { id } });
 
-    return new AdvertEntity({
-      ...advert,
-      owner: new OwnerDetailsEntity({
-        id: advert.owner.id,
-        username: advert.owner.username,
-        postalCode: advert.owner.postalCode,
-        createdAt: advert.owner.createdAt,
-        advertsCount: advert.owner._count.adverts,
-      }),
-      photos: advert.photos.map((photo) => new PhotoEntity(photo)),
+    return this.mapper.toAdvertDetailsEntity(advert, {
       isOwner: true,
       isFavorite: false,
-      favoriteCount: advert._count.favorites,
     });
   }
 
@@ -517,285 +414,5 @@ export class AdvertsService {
     }));
 
     return formattedCategories;
-  }
-
-  /**
-   * Adds photos to an advert owned by the authenticated user.
-   * @param advertId - Advert ID.
-   * @param userId - ID of the authenticated user.
-   * @param photoPaths - Array of photo URLs to add.
-   * @returns Updated AdvertEntity.
-   * @throws UnauthorizedException if user is not the owner.
-   */
-  async uploadPhoto(advertId: string, userId: string, photoPaths: string[]) {
-    const advert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-    });
-
-    assertOwnership(advert, userId);
-
-    await this.prisma.photo.createMany({
-      data: photoPaths.map((url) => ({ url, advertId })),
-    });
-
-    const updatedAdvert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: {
-        owner: {
-          select: {
-            id: true,
-            username: true,
-            postalCode: true,
-            createdAt: true,
-            _count: { select: { adverts: true } },
-          },
-        },
-        photos: true,
-        _count: { select: { favorites: true } },
-      },
-    });
-
-    return new AdvertEntity({
-      ...updatedAdvert,
-      owner: new OwnerDetailsEntity({
-        id: updatedAdvert.owner.id,
-        username: updatedAdvert.owner.username,
-        postalCode: updatedAdvert.owner.postalCode,
-        createdAt: updatedAdvert.owner.createdAt,
-        advertsCount: updatedAdvert.owner._count.adverts,
-      }),
-      photos: updatedAdvert.photos.map((photo) => new PhotoEntity(photo)),
-      isOwner: true,
-      isFavorite: false,
-      favoriteCount: updatedAdvert._count.favorites,
-    });
-  }
-
-  /**
-   * Removes a photo from an advert owned by the authenticated user.
-   * @param advertId - Advert ID.
-   * @param photoId - Photo ID to remove.
-   * @param userId - ID of the authenticated user.
-   * @returns Updated AdvertEntity.
-   * @throws UnauthorizedException if user is not the owner.
-   */
-  async deletePhoto(advertId: string, photoId: string, userId: string) {
-    const advert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: { photos: true },
-    });
-
-    assertOwnership(advert, userId);
-
-    const photo = advert.photos.find((p) => p.id === photo?.id);
-
-    if (!photo) {
-      throw new BadRequestException('This photo does not belong to the advert');
-    }
-
-    await this.prisma.photo.delete({ where: { id: photoId } });
-
-    const updatedAdvert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: {
-        owner: {
-          select: {
-            id: true,
-            username: true,
-            postalCode: true,
-            createdAt: true,
-            _count: { select: { adverts: true } },
-          },
-        },
-        photos: true,
-        _count: { select: { favorites: true } },
-      },
-    });
-
-    return new AdvertEntity({
-      ...updatedAdvert,
-      owner: new OwnerDetailsEntity({
-        id: updatedAdvert.owner.id,
-        username: updatedAdvert.owner.username,
-        postalCode: updatedAdvert.owner.postalCode,
-        createdAt: updatedAdvert.owner.createdAt,
-        advertsCount: updatedAdvert.owner._count.adverts,
-      }),
-      photos: updatedAdvert.photos.map((photo) => new PhotoEntity(photo)),
-      isOwner: true,
-      isFavorite: false,
-      favoriteCount: updatedAdvert._count.favorites,
-    });
-  }
-
-  /**
-   * Retrieves all adverts marked as favorite by the authenticated user.
-   * @param userId - ID of the authenticated user (nullable).
-   * @returns Array of AdvertEntity objects.
-   * @throws UnauthorizedException if user is not authenticated.
-   */
-  async getFavorites(userId: string) {
-    const favorites = await this.prisma.favorite.findMany({
-      where: { userId },
-      include: {
-        advert: {
-          include: {
-            owner: { select: { id: true, username: true } },
-            photos: true,
-            _count: { select: { favorites: true } },
-          },
-        },
-      },
-      orderBy: { createdAt: 'desc' },
-    });
-
-    return favorites.map(
-      (favorite) =>
-        new AdvertEntity({
-          ...favorite.advert,
-          owner: new OwnerSummaryEntity({
-            id: favorite.advert.owner.id,
-            username: favorite.advert.owner.username,
-          }),
-          photos: favorite.advert.photos.map((photo) => new PhotoEntity(photo)),
-          isOwner: false,
-          isFavorite: true,
-          favoriteCount: favorite.advert._count.favorites,
-          distance: null,
-        }),
-    );
-  }
-
-  /**
-   * Marks an advert as favorite for the authenticated user.
-   * @param advertId - Advert ID.
-   * @param userId - ID of the authenticated user.
-   * @returns Updated AdvertEntity with `isFavorite` set to `true`.
-   * @throws ForbiddenException if user owns the advert or already marked it as favorite.
-   */
-  async addFavorite(advertId: string, userId: string) {
-    const advert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: {
-        owner: {
-          select: {
-            id: true,
-            username: true,
-            postalCode: true,
-            createdAt: true,
-            _count: { select: { adverts: true } },
-          },
-        },
-        photos: true,
-        _count: { select: { favorites: true } },
-      },
-    });
-
-    if (advert.ownerId === userId) {
-      throw new ForbiddenException(
-        'Owned adverts cannot be marked as favorite',
-      );
-    }
-
-    const favoriteAdvert = await this.prisma.favorite.findUnique({
-      where: { userId_advertId: { userId, advertId } },
-    });
-
-    if (favoriteAdvert) {
-      throw new ForbiddenException('This advert is already marked as favorite');
-    }
-
-    await this.prisma.favorite.create({ data: { userId, advertId } });
-
-    const updatedAdvert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: {
-        owner: {
-          select: {
-            id: true,
-            username: true,
-            postalCode: true,
-            createdAt: true,
-            _count: { select: { adverts: true } },
-          },
-        },
-        photos: true,
-        _count: { select: { favorites: true } },
-      },
-    });
-
-    return new AdvertEntity({
-      ...updatedAdvert,
-      owner: new OwnerDetailsEntity({
-        id: updatedAdvert.owner.id,
-        username: updatedAdvert.owner.username,
-        postalCode: updatedAdvert.owner.postalCode,
-        createdAt: updatedAdvert.owner.createdAt,
-        advertsCount: updatedAdvert.owner._count.adverts,
-      }),
-      photos: updatedAdvert.photos.map((photo) => new PhotoEntity(photo)),
-      isOwner: false,
-      isFavorite: true,
-      favoriteCount: updatedAdvert._count.favorites,
-    });
-  }
-
-  /**
-   * Removes an advert marked as favorite from the authenticated user.
-   * @param advertId - Advert ID.
-   * @param userId - ID of the authenticated user.
-   * @returns Updated AdvertEntity with `isFavorite` set to `false`.
-   * @throws UnauthorizedException if user is not authenticated.
-   */
-  async deleteFavorite(advertId: string, userId: string) {
-    const favorite = await this.prisma.favorite.findUnique({
-      where: { userId_advertId: { userId, advertId } },
-    });
-
-    if (!favorite) {
-      throw new BadRequestException('This advert was not marked as favorite');
-    }
-
-    if (favorite.userId !== userId) {
-      throw new UnauthorizedException(
-        'You are not authorized to modify this advert',
-      );
-    }
-
-    await this.prisma.favorite.delete({
-      where: { userId_advertId: { userId, advertId } },
-    });
-
-    const updatedAdvert = await this.prisma.advert.findUniqueOrThrow({
-      where: { id: advertId },
-      include: {
-        owner: {
-          select: {
-            id: true,
-            username: true,
-            postalCode: true,
-            createdAt: true,
-            _count: { select: { adverts: true } },
-          },
-        },
-        photos: true,
-        _count: { select: { favorites: true } },
-      },
-    });
-
-    return new AdvertEntity({
-      ...updatedAdvert,
-      owner: new OwnerDetailsEntity({
-        id: updatedAdvert.owner.id,
-        username: updatedAdvert.owner.username,
-        postalCode: updatedAdvert.owner.postalCode,
-        createdAt: updatedAdvert.owner.createdAt,
-        advertsCount: updatedAdvert.owner._count.adverts,
-      }),
-      photos: updatedAdvert.photos.map((photo) => new PhotoEntity(photo)),
-      isOwner: false,
-      isFavorite: false,
-      favoriteCount: updatedAdvert._count.favorites,
-    });
   }
 }
